validador/requirements.txt
aniso8601==9.0.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==2.3.3
flask-marshmallow==0.15.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
requests==2.31.0
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.1
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
